requirements.txt
uvicorn
openai
fastapi
line-bot-sdk
python-dotenv